+++ python/lvmnps/__init__.py
@@ -14,10 +14,6 @@ NAME = "sdss-lvmnps"
 
 # Loads config. config name is the package name.
 config = get_config("lvmnps")
-
-# Inits the logging system as NAME. Only shell logging, and exception and warning catching.
-# File logging can be started by calling log.start_file_logger(path).  Filename can be different
-# than NAME.
 log = get_logger(NAME)
 
 # package name should be pip package name