+++ python/lvmnps/actor/commands/__init__.py
@@ -3,6 +3,7 @@ import importlib
 import os
 
 import click
+
 from clu.command import Command
 from clu.parsers.click import CluGroup, command_parser, help_, ping, version
 
@@ -22,7 +23,8 @@ parser.add_command(help_)
 def __commands(ctx, command: Command, *args):
     # Returns all commands.
 
-    # we have to use the help key for the command list, dont want to change the standard model.
+    # we have to use the help key for the command list,
+    # don't want to change the standard model.
     command.finish(help=[k for k in ctx.command.commands.keys() if k[:2] != "__"])