+++ python/lvmnps/actor/commands/status.py
@@ -9,6 +9,7 @@
 from __future__ import annotations
 
 import click
+
 from clu.command import Command
 
 from lvmnps.actor.commands import parser
@@ -28,7 +29,8 @@ async def status(
     elif switchname:
         if portnum:
             command.info(
-                text=f"Printing the current status of switch {switchname}, port {portnum}"
+                text="Printing the current status of switch "
+                f"{switchname}, port {portnum}"
             )
         else:
             command.info(text=f"Printing the current status of switch {switchname}")
@@ -41,7 +43,7 @@ async def status(
                 status[switch.name] = current_status
     elif switchname:
         for switch in switches:
-            # status |= await switch.statusAsDict(name, portnum) works only with python 3.9
+            # status |= await switch.statusAsDict(name, portnum) works only with PY 3.9
             if switchname == switch.name:
                 if portnum:
                     current_status = await switch.statusAsDict(switchname, portnum)